requirements.txt
Django==3.2.24
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
gunicorn==21.2.0
django-debug-toolbar==4.3.0
ipython==8.22.1